backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
motor==3.7.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.6
spotipy==2.23.0
yt-dlp==2025.6.30
ytmusicapi==1.10.2
google-api-python-client==2.165.0
